READMe.md
The .html file is the prototype of a closed universe contagion model. The Microsoft Word document gives instructions for its execution and operation with various parameter settings. In particular, one can see how different intensities of epidemics can be generated by selected settings, and reduced dramatically by interventions through changing parameters. At this point, this is an analytic model, not a predictive model.